requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
pyyaml>=6.0
